java Code Convention에 맞춰 수정함
